+++ src/abstractdataframe/reshape.jl
@@ -198,11 +198,11 @@ end
 
 """
     unstack(df::AbstractDataFrame, rowkeys, colkey, value; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
     unstack(df::AbstractDataFrame, colkey, value; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
     unstack(df::AbstractDataFrame; renamecols::Function=identity,
-            allowmissing::Bool=false, allowduplicates::Bool=false)
+            allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
 
 Unstack data frame `df`, i.e. convert it from long to wide format.
 
@@ -229,6 +229,8 @@ Row and column keys will be ordered in the order of their first appearance.
 - `allowduplicates`: if `false` (the default) then an error an error will be thrown
   if combination of `rowkeys` and `colkey` contains duplicate entries; if `true`
   then  then the last encountered `value` will be retained.
+- `fillvalue`: missing row/column combinations are filled with this value; The default
+  is `missing`.
 
 # Examples
 
@@ -331,12 +333,33 @@ julia> unstack(long, :id, :variable, :value, renamecols=x->Symbol(:_, x))
    4 │     4       2.0       1.0       2.0
    5 │     5       2.0       1.0       3.0
    6 │     6       2.0       1.0       3.0
+
+julia> df = DataFrame(
+    :id => ["1", "1", "2"],
+    :variable => ["Var1", "Var2", "Var1"],
+    :value => [1, 2, 3]
+)
+3×3 DataFrame
+ Row │ id      variable  value 
+     │ String  String    Int64 
+─────┼─────────────────────────
+   1 │ 1       Var1          1
+   2 │ 1       Var2          2
+   3 │ 2       Var1          3
+
+julia> unstack(df, :variable, :value, fillvalue=42)
+2×3 DataFrame
+ Row │ id      Var1   Var2  
+     │ String  Int64  Int64 
+─────┼──────────────────────
+   1 │ 1           1      2
+   2 │ 2           3     42
 ```
 Note that there are some differences between the widened results above.
 """
 function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
                  value::ColumnIndex; renamecols::Function=identity,
-                 allowmissing::Bool=false, allowduplicates::Bool=false)
+                 allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
     rowkey_ints = vcat(index(df)[rowkeys])
     @assert rowkey_ints isa AbstractVector{Int}
     length(rowkey_ints) == 0 && throw(ArgumentError("No key column found"))
@@ -344,23 +367,23 @@ function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
     g_colkey = groupby(df, colkey)
     valuecol = df[!, value]
     return _unstack(df, rowkey_ints, index(df)[colkey], g_colkey,
-                    valuecol, g_rowkey, renamecols, allowmissing, allowduplicates)
+                    valuecol, g_rowkey, renamecols, allowmissing, allowduplicates, fillvalue)
 end
 
 function unstack(df::AbstractDataFrame, colkey::ColumnIndex, value::ColumnIndex;
                  renamecols::Function=identity,
-                 allowmissing::Bool=false, allowduplicates::Bool=false)
+                 allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing)
     colkey_int = index(df)[colkey]
     value_int = index(df)[value]
     return unstack(df, Not(colkey_int, value_int), colkey_int, value_int,
             renamecols=renamecols, allowmissing=allowmissing,
-            allowduplicates=allowduplicates)
+            allowduplicates=allowduplicates, fillvalue=fillvalue)
 end
 
 unstack(df::AbstractDataFrame; renamecols::Function=identity,
-        allowmissing::Bool=false, allowduplicates::Bool=false) =
+        allowmissing::Bool=false, allowduplicates::Bool=false, fillvalue=missing) =
     unstack(df, :variable, :value, renamecols=renamecols, allowmissing=allowmissing,
-            allowduplicates=allowduplicates)
+            allowduplicates=allowduplicates, fillvalue=fillvalue)
 
 # we take into account the fact that idx, starts and ends are computed lazily
 # so we rather directly reference the gdf.groups
@@ -388,7 +411,7 @@ function _unstack(df::AbstractDataFrame, rowkeys::AbstractVector{Int},
                   colkey::Int, g_colkey::GroupedDataFrame,
                   valuecol::AbstractVector, g_rowkey::GroupedDataFrame,
                   renamecols::Function,
-                  allowmissing::Bool, allowduplicates::Bool)
+                  allowmissing::Bool, allowduplicates::Bool, fillvalue)
     rowref = g_rowkey.groups
     row_group_row_idxs = find_group_row(g_rowkey)
     Nrow = length(g_rowkey)
@@ -398,13 +421,15 @@ function _unstack(df::AbstractDataFrame, rowkeys::AbstractVector{Int},
     Ncol = length(g_colkey)
     col_group_row_idxs = find_group_row(g_colkey)
     colref_map = df[col_group_row_idxs, colkey]
-
     if any(ismissing, colref_map) && !allowmissing
         throw(ArgumentError("Missing value in variable :$(_names(df)[colkey]). " *
                             "Pass `allowmissing=true` to skip missings."))
     end
+    unstacked_val = [fill!(
+        similar(valuecol, promote_type(eltype(valuecol), typeof(fillvalue)), Nrow),
+        fillvalue
+    ) for _ in 1:Ncol]
 
-    unstacked_val = [similar_missing(valuecol, Nrow) for i in 1:Ncol]
     mask_filled = falses(Nrow, Ncol)
 
     @assert length(rowref) == length(colref) == length(valuecol)